fix computation of system matrix and integration matrix rhs
Move the construction of the matrices from `initialization` to the beginning of `advanceTimeStep`, because the time step width depends on the integration interval, which is not known until the integration start.
Without this fix, a wrong time step size was used, especially if the integrator was based on 'number of steps per interval'.
Therefore we initialize the matrices with the correct time step width in the first call to `advanceTimeStep`. The matrices are only recomputed if the time step changes.

This, also fixes the problem, that different time step sizes lead to different results with 'multiple fibers' (and probably all other simulations, too).