fix time stepping scheme and CellML with paraview writer
This affects all simulations that perform time steps after the first paraview output.
Both, computed values that are used later in the simulation and the written values are wrong after the first time the output writer was called.
Simulations that switching to another solver directly after the output is written are unaffected. (each solver resets the representation of the field variables.

The time stepping schemes assume that the field variables are in contiguous representation. However, the paraview writer has to change the representation to communicate ghost values.
Now, the paraview writer resets the representation of the field variables to their original represenatation after the communication is done.

Copying back the values to the original representation is avoided because the writer only _reads_ the data.